+++ p2panda-discovery/README.md
@@ -1,0 +1,1 @@
+Awesome documentation goes here.
